[Refactor] Add schema change utils
Some common codes are moved to utils and will be used in lake table schema change.
